Update data.json
Changes made to intruders in black dont work,
because in the data json the quest wasnt tagged as  "questRewardsSkills": true
therefore there is a bug that some gems wont be shown properly in the editor.
I need to push this update with the risk of overriding user's custom data files for zones.